+++ R/list_cancer_type.R
@@ -62,4 +62,7 @@ list_cancer_subtypes <- function(input_disease, data_dir) {
     dplyr::filter(.data$disease %in% input_disease) %>%
     dplyr::pull(.data$disease_subtype) %>%
     unique
+  
+  # Message:
+  message("See sample_annot.rda in your data directory for a complist list of cell line annotation.")
 }